WalkingkookaLocale.isUnsupported taken from WalkingkookaLocale
- Part of effort to extra as much as possible from WalkingkookaLocale.
